pyproject.toml
[tool.poetry]
name = "repl-nix-medicalexammanager"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
openai = "^1.45.0"
plotly = "^5.24.1"
scikit-learn = "^1.5.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
